feat: pre-commit, clang-tidy config, update clang-format config for json, remove fmt dependency, use clang++ on unix, format all files, apply a lot of linting suggestions

+++ src/gui/gui.h
@@ -9,7 +9,7 @@ namespace gui {
 	const inline float vsync_extra_fps = 50;
 	const inline float min_delta_time = 1.f / 10;
 	const inline float default_delta_time = 1.f / 60;
-	inline float vsync_frame_time = default_delta_time;
+	inline double vsync_frame_time = default_delta_time;
 
 	void update_vsync();
